[LibOS] Change ioctl(, int cmd) to ioctl(, unsigned long cmd)
The second argument to ioctl is unsigned long, not int.

+++ LibOS/shim/src/sys/shim_ioctl.c
@@ -303,7 +303,7 @@ void signal_io(IDTYPE target, void* arg) {
     put_thread(thread);
 }
 
-int shim_do_ioctl(int fd, int cmd, unsigned long arg) {
+int shim_do_ioctl(int fd, unsigned long cmd, unsigned long arg) {
     struct shim_handle* hdl = get_fd_handle(fd, NULL, NULL);
     if (!hdl)
         return -EBADF;
